Price data contract ticket (#2)
* Push for ticket

* Updates to remove test errors and warnings

* Updates to remove useless cargo imports

* Updates to remove useless cargo imports

* All test cases passing

* All warnings removed

* Addressed comments

+++ src/data/price_posting.rs
@@ -1,3 +1,4 @@
+use cosmwasm_std::Decimal;
 use getset::Getters;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -7,12 +8,12 @@ use sp_secret_toolkit::macros::{identifiable::Identifiable, keymap::KeymapStorag
     Getters, Serialize, Deserialize, Clone, Debug, Eq, PartialEq, JsonSchema, KeymapStorage,
 )]
 pub struct PricePosting {
-    price: u32,
+    price: Decimal,
     time: u64,
 }
 
 impl PricePosting {
-    pub fn new(price: u32, time: u64) -> Self {
+    pub fn new(price: Decimal, time: u64) -> Self {
         PricePosting { price, time }
     }
 }
